third_party/WebKit/LayoutTests/fast/canvas-api/canvas-ellipse-negative-radius.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {
    var canvas = document.createElement('canvas');
    var ctx = canvas.getContext('2d');

    ctx.ellipse(10, 10, 10, 5, 0, 0, 1, false);
    ctx.ellipse(10, 10, 10, 0, 0, 0, 1, false);
    ctx.ellipse(10, 10, -0, 5, 0, 0, 1, false);

    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.ellipse(10, 10, -2, 5, 0, 0, 1, false);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.ellipse(10, 10, 0, -1.5, 0, 0, 1, false);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.ellipse(10, 10, -2, -5, 0, 0, 1, false);
    });
    // FIXME: When this is exposed on Path as well as CanvasRenderingContext2D, add tests for Path as well.
}, 'Tests CanvasPath ellipse with negative radii.');
</script>
</body>
